move cache and test import types as parameters
