+++ native/src/open_api/generator.rs
@@ -1,13 +1,9 @@
-use std::collections::BTreeMap;
-
-use ahash::{HashMap, HashMapExt, HashSet, HashSetExt};
+use ahash::{HashSet, HashSetExt};
 use neon::prelude::FunctionContext;
 
 use crate::typescript::*;
 
 use super::open_api::{ApiPathOperation, ApiSchema, OpenApi, PathArgs, ResponseOptions};
-
-static mut DECLARATIONS: BTreeMap<String, HashMap<String, Declaration>> = BTreeMap::new();
 
 fn add_body_parameter(
     operation: &mut ApiPathOperation,
@@ -175,167 +171,6 @@ fn update_schema(schema: &mut ApiSchema, node: &AstNode) -> () {
     }
 }
 
-fn cache_declarations(node: &AstNode, file_name: &str) -> () {
-    match node.kind {
-        CLASS_DECLARATION => cache_object_type(node, file_name),
-        IMPORT_DECLARATION => cache_import_declaration(node, file_name),
-        EXPORT_DECLARATION => cache_export_declaration(node, file_name),
-        INTERFACE_DECLARATION => cache_object_type(node, file_name),
-        TYPE_ALIAS_DECLARATION => cache_object_type(node, file_name),
-        VARIABLE_STATEMENT => cache_variables(node, file_name),
-        _ => node.for_each_child(|n| cache_declarations(n, file_name)),
-        // _ => {}
-    }
-}
-
-fn cache_object_type(node: &AstNode, file_name: &str) -> () {
-    unsafe {
-        let name = node.name.as_ref().unwrap();
-        let text = match node.modifiers {
-            Some(ref modifiers) => match modifiers.iter().find(|n| n.kind == DEFAULT_KEYWORD) {
-                Some(_) => "default",
-                None => name.escaped_text.as_ref().unwrap(),
-            },
-            None => name.escaped_text.as_ref().unwrap(),
-        };
-
-        let declarations = DECLARATIONS.entry(file_name.to_string()).or_insert(HashMap::new());
-        declarations.insert(text.to_string(), Declaration::Type { node: node.clone() });
-    }
-}
-
-fn cache_import_declaration(node: &AstNode, file_name: &str) -> () {
-    unsafe {
-        let declarations = DECLARATIONS.entry(file_name.to_owned()).or_insert(HashMap::new());
-        let module_specifier = node.module_specifier.as_ref().unwrap();
-        let module_reference = module_specifier.text.as_ref().unwrap();
-
-        let import_clause = node.import_clause.as_ref().unwrap();
-        if let Some(ref name) = import_clause.name {
-            let text = name.escaped_text.as_ref().unwrap();
-            declarations.insert(
-                text.to_string(),
-                Declaration::Import {
-                    name: String::from("default"),
-                    module_ref: module_reference.to_string(),
-                },
-            );
-        }
-
-        import_clause.for_each_child(|bindings| {
-            bindings.for_each_child(|element| {
-                let name = element.name.as_ref().unwrap();
-                let name_text = name.escaped_text.as_ref().unwrap();
-                let alias = match element.property_name {
-                    Some(ref node) => node.escaped_text.as_ref().unwrap(),
-                    None => name_text,
-                };
-
-                declarations.insert(
-                    name_text.to_string(),
-                    Declaration::Import {
-                        name: alias.to_string(),
-                        module_ref: module_reference.to_string(),
-                    },
-                );
-            })
-        })
-    }
-}
-
-fn cache_export_declaration(node: &AstNode, file_name: &str) -> () {
-    unsafe {
-        let declarations = DECLARATIONS.entry(file_name.to_owned()).or_insert(HashMap::new());
-        let module_specifier = node.module_specifier.as_ref().unwrap();
-        let module_reference = module_specifier.text.as_ref().unwrap();
-
-        let export_clause = node.export_clause.as_ref().unwrap();
-
-        export_clause.for_each_child(|exports| {
-            exports.for_each_child(|element| {
-                let name = element.name.as_ref().unwrap();
-                let name_text = name.escaped_text.as_ref().unwrap();
-                let alias = match element.property_name {
-                    Some(ref node) => node.escaped_text.as_ref().unwrap(),
-                    None => name_text,
-                };
-
-                declarations.insert(
-                    name_text.to_string(),
-                    Declaration::Export {
-                        name: alias.to_string(),
-                        module_ref: module_reference.to_string(),
-                    },
-                );
-            })
-        })
-    }
-}
-
-fn cache_variables(node: &AstNode, file_name: &str) -> () {
-    unsafe {
-        let declarations = DECLARATIONS.get_mut(file_name).unwrap();
-        if let Some(ref list) = node.declaration_list {
-            list.for_each_child(|declaration| {
-                let initializer = declaration.initializer.as_ref().unwrap();
-                let name = declaration.name.as_ref().unwrap();
-                let text = name.escaped_text.as_ref().unwrap();
-                match initializer.kind {
-                    AS_EXPRESSION => {
-                        let _type = initializer._type.as_ref().unwrap();
-                        let type_name = _type.type_name.as_ref().unwrap();
-                        let type_name_text = type_name.escaped_text.as_ref().unwrap();
-                        declarations.insert(
-                            text.to_string(),
-                            Declaration::Alias {
-                                from: text.to_string(),
-                                to: type_name_text.to_string(),
-                            },
-                        );
-                    }
-                    TYPE_ASSERTION_EXPRESSION => {
-                        let _type = initializer._type.as_ref().unwrap();
-                        let type_name = _type.type_name.as_ref().unwrap();
-                        let type_name_text = type_name.escaped_text.as_ref().unwrap();
-                        declarations.insert(
-                            text.to_string(),
-                            Declaration::Alias {
-                                from: text.to_string(),
-                                to: type_name_text.to_string(),
-                            },
-                        );
-                    }
-                    CALL_EXPRESSION => {
-                        let expression = initializer.expression.as_ref().unwrap();
-                        let expression_text = expression.escaped_text.as_ref().unwrap();
-                        declarations.insert(
-                            text.to_string(),
-                            Declaration::Alias {
-                                from: text.to_string(),
-                                to: expression_text.to_string(),
-                            },
-                        );
-                    }
-                    NEW_EXPRESSION => {
-                        let expression = initializer.expression.as_ref().unwrap();
-                        let expression_text = expression.escaped_text.as_ref().unwrap();
-                        declarations.insert(
-                            text.to_string(),
-                            Declaration::Alias {
-                                from: text.to_string(),
-                                to: expression_text.to_string(),
-                            },
-                        );
-                    }
-                    _ => {
-                        declarations.insert(text.to_string(), Declaration::Type { node: node.clone() });
-                    }
-                }
-            })
-        }
-    }
-}
-
 fn find_type_name(name: &str, file_name: &String) -> String {
     unsafe {
         let mut key = name;
@@ -355,7 +190,7 @@ fn find_type_name(name: &str, file_name: &String) -> String {
     }
 }
 
-fn get_declaration<'n, F>(
+pub fn get_declaration<'n, F>(
     reference: &str,
     module_ref: &str,
     src_file_name: &String,
@@ -571,22 +406,23 @@ impl<F: FnMut(&str, &str, &mut FunctionContext) -> AstNode> OpenApiGenerator<F> 
         }
     }
 
-    fn find_api_paths(&mut self, node: &AstNode, file_name: &String) -> () {
+    fn find_api_paths(&mut self, node: &AstNode, file_name: &String, cx: &mut FunctionContext) -> () {
         if self.is_api_path(node) {
             let arguments = node.arguments.as_ref().unwrap();
             let route_handler = arguments.get(0).unwrap();
-            let route_handler_body = route_handler.body.as_ref().unwrap();
-            let request_param = get_request_parameter(route_handler);
-            let path_options = get_path_args(arguments.get(1).unwrap());
 
+            let path_options = get_path_args(arguments.get(1).unwrap());
             let route = path_options.path.unwrap();
             let method = path_options.method.unwrap();
             self.open_api.path(&route).method(&method).tags(path_options.tags);
 
+            let request_param = get_request_parameter(route_handler);
             self.add_operation_params(&route, &method, request_param, file_name);
+
+            let route_handler_body = route_handler.body.as_ref().unwrap();
             self.add_operation_response(&route, &method, &*route_handler_body, file_name);
         } else {
-            node.for_each_child(|node| self.find_api_paths(node, file_name))
+            node.for_each_child(|node| self.find_api_paths(node, file_name, cx))
         }
     }
 
@@ -595,7 +431,7 @@ impl<F: FnMut(&str, &str, &mut FunctionContext) -> AstNode> OpenApiGenerator<F> 
 
         cache_declarations(&source_file, &path);
 
-        source_file.for_each_child(|node| self.find_api_paths(node, &path));
+        source_file.for_each_child(|node| self.find_api_paths(node, &path, cx));
 
         let mut references = self.references.iter().collect::<Vec<&TypeReference>>();
         while references.is_empty() == false {
