+++ native/src/typescript/mod.rs
@@ -1,5 +1,7 @@
+mod cache;
 mod source_file;
 mod syntax_kind;
 
+pub use self::cache::*;
 pub use self::source_file::*;
 pub use self::syntax_kind::*;
