+++ tests/test-api/dtos/index.ts
@@ -1,4 +1,6 @@
-export { default as UserPatch, CreateUserRequest } from './requests'
+import { Header, QueryParam } from '../../../src';
+
+export { CreateUserRequest, default as UserPatch } from './requests';
 
 export interface User {
     name: string;
@@ -8,4 +10,6 @@ export default class AdminUser implements User {
     permissions!: string[];
     name!: string;
 }
+
+export interface GetUserRequest { lat: QueryParam<number, false>, long: QueryParam<number, false>; headers: { user: Header<User, true, "v1">; }; }
